+++ dotnet/src/GameBoyAdvance/GbaGsFcdRom.cs
@@ -1,5 +1,5 @@
 using System.Globalization;
-using LibreShark.Hammerhead.Nintendo64;
+using LibreShark.Hammerhead.Codecs;
 using LibreShark.Hammerhead.IO;
 
 namespace LibreShark.Hammerhead.GameBoyAdvance;
