+++ dotnet/src/GameBoyAdvance/GbaTvTunerRom.cs
@@ -1,4 +1,4 @@
-using LibreShark.Hammerhead.Nintendo64;
+using LibreShark.Hammerhead.Codecs;
 using LibreShark.Hammerhead.IO;
 
 namespace LibreShark.Hammerhead.GameBoyAdvance;
