+++ dotnet/src/GameBoyColor/GbcGsV3Rom.cs
@@ -1,5 +1,5 @@
 using Google.Protobuf;
-using LibreShark.Hammerhead.Nintendo64;
+using LibreShark.Hammerhead.Codecs;
 using LibreShark.Hammerhead.IO;
 
 namespace LibreShark.Hammerhead.GameBoyColor;
