+++ dotnet/src/Nintendo64/N64GbHunterRom.cs
@@ -1,3 +1,4 @@
+using LibreShark.Hammerhead.Codecs;
 using LibreShark.Hammerhead.IO;
 using Spectre.Console;
