+++ dotnet/src/Program.cs
@@ -3,8 +3,8 @@ using System.Reflection;
 using System.Text.RegularExpressions;
 using Google.Protobuf;
 using Google.Protobuf.Collections;
+using LibreShark.Hammerhead.Codecs;
 using LibreShark.Hammerhead.IO;
-using LibreShark.Hammerhead.Nintendo64;
 
 namespace LibreShark.Hammerhead;
